Adopt SwiftLint code style
This leaves a few extra @TODO statements in the code where errors might need to be handled in the future.

While this makes some code a bit more verbose, it also handles a few cases which would soon become a problem (ie. if we have different UITableViewCells for the Library, etc)

Re