Feature: Add support section to settings
Includes DeviceKit to get a human readable hardware identifier from the hardware version. Adds email and GitHub as possible support options, making it as easy as possible to send actionable problemms.

The email snippet contains the version of BookPlayer, the device and the OS version as well as a writing promt. We could consider extending this to include the usual bug report taglines in the future but until we get the first emails I'd stay with this simple version.

If no email account is available the info and email can be copied to the clipboard from the alert dialog.

This also includes a new run script which increases the build number automatically.

Re #59