Bugfix: Ensure enough colors are extracted from an artwork
This also abstracts handling for an empty artwork. If no colors are found, the default look will be loaded (nb: the defaults I set are super boring, I'll have to find something better here).

If not enough colors are extracted, the missing colors will be supplemented by either black or white depending the base color

Re #122