Feature: Hint cover drag for first time users
This enables a drag hint that nudges the artwork every time the player is presented until the user uses the drag gesture to rewind or forward.

The same nudge is used when the icons behind the artwork are tapped in an attempt to jump that way.

This should provide enough feedback and encouragement to users that return from an earlier version and also convey a little achievement when they interact with the artwork the first time.