Test dismissal for jellyfin views when download starts
Also adds MockCoordinatorPresentationFlow to help testing a flow getting dismissed:

MockNavigationController didn't work here.
The JellyfinCoordinator usually uses a Modal flow. When it is supposed to be hidden, we call flow.finishPresentation.
However, when used with pushFlow and MockNavigationController for the unit tests, this will only popToRootViewController, leaving us with a controller on the horizontal stack.

I also cannot switch to modalFlow, because then I do not have a MockNavigationController for checking the pushed view controllers.

An alternative would be extending the modalFlow to accept a factory for the navigation controller (which could then me a MockNavigationController).
But that's pretty much what MockCoordinatorPresentationFlow is.