Don't smart rewind past beginning of chapter
Currently, smart rewind will always put you back 30 seconds, even if you
are very close to the start of a chapter. This is probably not the
behavior that most users would expect or desire. So instead we can limit
rewinding so that we never cross backwards over a chapter boundary.