Add smart rewind interval setting, improve easing
This adds a new setting that allows users to speficy the maximum amount
of time they want the smart rewind feature to skip backwards by.

In addition, it changes how the the smart rewind feature calculates the
amount of time to skip backwards:

It adds a minimum rewind interval of 2 seconds, so that when smart
rewind is enabled, playback will always rewind by at least 2 seconds.
The only exceptions are if we have been paused for less than 2 seconds
or if we are less than 2 seconds from the start of the chapter. In these
cases we will skip backwards by either the amount of time paused, or to
the beginning of the chapter, whichever is closest.

It changes the easing function to be ease-out instead of ease-in.
Instead of an easing function that starts slowly, then quickly increases
towards the end I think the opposite is more natural. The amount we
forget over time quickly increases at first, then slows down as time
goes on. So I think the easing function should follow this. Also IMHO
quartic feels like a little bit better of a fit than cubic here.

Finally, this commit changes smart rewind threshold up from 10 to 60
minutes. This seems like a big change, but with the change of the easing
function it's less extreme. IMO 10 minutes was a bit too short of a
threshold, as personally I would want to skip back further after e.g. 45
minutes of pause time than after 10. See attached graphs to get a feel
for how this looks in practice.